requirements.txt
torch
huggingface_hub
transformers
geoopt
tqdm
numpy
einops
datasets
scipy
wandb
accelerate
pyyaml